Notes can now be created, edited and deleted.
MUI, aka Material-UI, was added to the project.
Added a loading component to the project.
Added a loading component to the notes page.
Added an error component to the notes page.
Added an editor component to the notes page.
Edited the layout component to include the MUI theme.
Started some basic styling of the notes page.